Move parseQuery method into another module
It's unintuitive if parseQuery is located in mqttDeserialize module.
