+++ client/src/map/mapContainerTest.tsx
@@ -23,12 +23,14 @@ describe('<MapContainer />', () => {
       <MemoryRouter initialEntries={['/map?lat=1&lon=2&host=abc&topic=aihe']}>
         <MapContainer
           isAdminPanelOpen={false}
+          bluetoothName="phone-1234"
+          setBluetoothName={() => {}}
+          nameModalOpen={false}
+          setNameModalOpen={() => {}}
           roomReserved={false}
           getDeviceLocation={null}
           setDeviceLocation={() => {}}
           devices={[]}
-          shareLocationModalOpen={false}
-          openShareLocationModal={() => {}}
         />
       </MemoryRouter>
     );
