+++ client/src/router.tsx
@@ -1,10 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import {
-  BrowserRouter,
-  Route,
-  Switch,
-  RouteComponentProps,
-} from 'react-router-dom';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import styled from 'styled-components';
 import { Transition } from 'react-spring/renderprops';
 import AboutContainer from './aboutContainer';
@@ -22,9 +17,7 @@ import LoginPromptContainer from './admin/loginPromptContainer';
 import AuthApi, { Admin } from './admin/authApi';
 import ShareLocationModal from './map/shareLocationModal';
 import PublicShareModal from './map/publicShareModal';
-import Deserializer, {
-  MapLocationQueryDecoder,
-} from './location/mqttDeserialize';
+import { parseQuery, MapLocationQueryDecoder } from './common/urlParse';
 
 const NotFound = () => <h3>404 page not found</h3>;
 
@@ -40,8 +33,6 @@ const MainRow = styled.div`
 `;
 
 const Router = () => {
-  const parser = new Deserializer();
-
   const [admin, setAdmin] = useState<Admin | null>(null);
   const [isAdminPanelOpen, openAdminPanel] = useState(false);
   const [getDeviceLocation, setDeviceLocation] = useState<Location | null>(
@@ -58,7 +49,7 @@ const Router = () => {
   const [publicShareOpen, openPublicShare] = useState(false);
   const [bluetoothName, setBluetoothName] = useState<string | null>(null);
 
-  const queryParams = parser.parseQuery(
+  const queryParams = parseQuery(
     MapLocationQueryDecoder,
     document.location.search
   );
